refactor: Moves LT authoring components to simplify organization and solves minor UI bug.
