+++ src/app/modules/authoring/authoring.module.ts
@@ -19,11 +19,11 @@ import { MrqFormComponent } from './assessment-items/mrq-form/mrq-form.component
 import { McqFormComponent } from './assessment-items/mcq-form/mcq-form.component';
 import { SaqFormComponent } from './assessment-items/saq-form/saq-form.component';
 import {LearningModule} from "../learning/learning.module";
-import { ActivitiesComponent } from './learning-tasks/activities/activities.component';
-import { ActivityDetailsComponent } from './learning-tasks/activities/activity-details/activity-details.component';
 import { LearningTasksComponent } from './learning-tasks/learning-tasks.component';
-import { LearningTaskFormComponent } from './learning-tasks/learning-task-form/learning-task-form.component';
+import { TaskCloningFormComponent } from './learning-tasks/task-cloning-form/task-cloning-form.component';
 import { LearningTaskComponent } from './learning-tasks/learning-task/learning-task.component';
+import { ActivitiesComponent } from './learning-tasks/learning-task/activities/activities.component';
+import { ActivityDetailsComponent } from './learning-tasks/learning-task/activity-details/activity-details.component';
 import { LearningTaskDetailsComponent } from './learning-tasks/learning-task/learning-task-details/learning-task-details.component'
 
 @NgModule({
@@ -43,7 +43,7 @@ import { LearningTaskDetailsComponent } from './learning-tasks/learning-task/lea
     ActivitiesComponent,
     ActivityDetailsComponent,
     LearningTasksComponent,
-    LearningTaskFormComponent,
+    TaskCloningFormComponent,
     LearningTaskComponent,
     LearningTaskDetailsComponent
   ],
