+++ src/app/modules/authoring/learning-tasks/learning-task/activities/activities.component.scss
