+++ src/app/modules/authoring/learning-tasks/learning-task/activity-details/activity-details.component.html
@@ -9,16 +9,16 @@
                 <mat-label>Naziv</mat-label>
                 <input formControlName="name" matInput [readonly]="!editMode">
             </mat-form-field>
-            <section fxLayout="row" fxFlex="15">
+            <section fxLayout="row" fxFlex="15" fxLayoutGap="10px">
                 <span fxFlex></span>
-                <button mat-icon-button *ngIf="!editMode" (click)="edit()">
+                <button mat-mini-fab *ngIf="!editMode" (click)="edit()">
                     <mat-icon matTooltip="Izmeni" [matTooltipPosition]="'above'">edit</mat-icon>
                 </button>
-                <button mat-icon-button *ngIf="editMode" color="primary" [disabled]="!activityForm.valid"
+                <button mat-mini-fab *ngIf="editMode" color="primary" [disabled]="!activityForm.valid"
                     (click)="submit()">
                     <mat-icon matTooltip="Sačuvaj" [matTooltipPosition]="'above'">check</mat-icon>
                 </button>
-                <button mat-icon-button *ngIf="editMode" color="warn" (click)="discardChanges()">
+                <button mat-mini-fab *ngIf="editMode" color="warn" (click)="discardChanges()">
                     <mat-icon matTooltip="Odbaci" [matTooltipPosition]="'above'">clear</mat-icon>
                 </button>
             </section>
@@ -39,11 +39,11 @@
                     <div [formGroup]="exampleForm" fxLayout="row" fxLayoutGap="10px" fxLayoutAlign=" center">
                         <mat-form-field [appearance]="editMode ? 'fill' : 'outline'" fxFlex="30">
                             <mat-label>Kod</mat-label>
-                            <input matInput formControlName="code">
+                            <input matInput formControlName="code" [readonly]="!editMode">
                         </mat-form-field>
                         <mat-form-field [appearance]="editMode ? 'fill' : 'outline'" fxFlex="70">
                             <mat-label>Url</mat-label>
-                            <input matInput formControlName="url">
+                            <input matInput formControlName="url" [readonly]="!editMode">
                         </mat-form-field>
                         <button *ngIf="editMode" mat-icon-button
                             (click)="deleteExample(i)"><mat-icon>delete_forever</mat-icon></button>
@@ -82,15 +82,15 @@
                     <div [formGroup]="standardForm" fxLayout="row" fxLayoutGap="10px" fxLayoutAlign=" center">
                         <mat-form-field fxFlex="30" [appearance]="editMode ? 'fill' : 'outline'">
                             <mat-label>Naziv</mat-label>
-                            <input matInput formControlName="name">
+                            <input matInput formControlName="name" [readonly]="!editMode">
                         </mat-form-field>
                         <mat-form-field fxFlex="60" [appearance]="editMode ? 'fill' : 'outline'">
                             <mat-label>Opis</mat-label>
-                            <input matInput formControlName="description">
+                            <input matInput formControlName="description" [readonly]="!editMode">
                         </mat-form-field>
                         <mat-form-field fxFlex="10" [appearance]="editMode ? 'fill' : 'outline'">
                             <mat-label>Bodovi</mat-label>
-                            <input matInput type="number" formControlName="maxPoints">
+                            <input matInput type="number" formControlName="maxPoints" [readonly]="!editMode">
                         </mat-form-field>
                         <button *ngIf="editMode" mat-icon-button
                             (click)="deleteStandard(i)"><mat-icon>delete_forever</mat-icon></button>
