+++ src/app/modules/authoring/learning-tasks/learning-task/activity-details/activity-details.component.scss
