+++ src/app/modules/authoring/learning-tasks/task-cloning-form/task-cloning-form.component.ts
@@ -5,15 +5,15 @@ import { LearningTask } from '../model/learning-task';
 import { Activity } from '../model/activity';
 
 @Component({
-  selector: 'cc-learning-task-form',
-  templateUrl: './learning-task-form.component.html',
-  styleUrls: ['./learning-task-form.component.scss']
+  selector: 'cc-task-cloning-form',
+  templateUrl: './task-cloning-form.component.html',
+  styleUrls: ['./task-cloning-form.component.scss']
 })
-export class LearningTaskFormComponent  {
+export class TaskCloningFormComponent  {
   learningTaskForm: FormGroup;
   template: LearningTask;
 
-  constructor(private builder: FormBuilder, private dialogref: MatDialogRef<LearningTaskFormComponent>, @Inject(MAT_DIALOG_DATA) public data: any) { 
+  constructor(private builder: FormBuilder, private dialogref: MatDialogRef<TaskCloningFormComponent>, @Inject(MAT_DIALOG_DATA) public data: any) { 
     this.template = data?.template;
     this.createForm();
   }
